requirements.txt
# Core Scientific Computing
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0

# Machine Learning & Deep Learning
scikit-learn>=0.24.0
tensorflow>=2.6.0
xgboost>=1.5.0

# Topological Data Analysis
ripser>=0.0.15
gudhi>=3.4.0

# Bioinformatics
biopython>=1.79

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0
networkx>=2.6.0
umap-learn>=0.5.0

# Utilities
tqdm>=4.60.0
joblib>=1.0.0
pillow>=8.3.0

# Python version requirement
# Requires Python 3.7+
